+++ server/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+### Added:
+
+- support running with buildout < 2.9.3
+
 ### Fixed
 
 - fixed broken v0.1.1 release, it could not be installed from pypi.